move endpoints to separate router (clientProgramRouter) and addDELETE endpoint
