+++ api/server.js
@@ -6,6 +6,7 @@ const exercisesRouter = require('../exercises/exercisesRouter');
 const workoutsRouter = require('../workouts/workoutsRouter');
 const clientsRouter = require('../clients/clientsRouter');
 const programsRouter = require('../programs/programsRouter');
+const clientProgramRouter = require('../clients/clientProgramRouter');
 
 // create server
 const server = express();
@@ -24,5 +25,6 @@ server.use('/exercises', exercisesRouter);
 server.use('/workouts', workoutsRouter);
 server.use('/clients', clientsRouter);
 server.use('/programs', programsRouter);
+server.use('/clients-programs', clientProgramRouter);
 
 module.exports = server;
