+++ clients/clientsRouter.js
@@ -107,50 +107,4 @@ router.put('/:id', validTokenCheck, validBodyCheck(['first_name', 'last_name', '
     });
 });
 
-// --------------------------- CLIENTS_PROGRAMS ---------------------------- //
-
-// ********************************************************
-// POST /clients/programs
-// ********************************************************
-router.post('/programs', validTokenCheck, (req, res) => {
-  let programId = req.body.id;
-  //const coach_id = req.token.coachID;
-  let clientProgramArray = req.body.clients.map(el => {
-    let eachObject = {};
-    eachObject.client_id = el.id;
-    eachObject.program_id = programId;
-    eachObject.start_date = el.start_date;
-    return eachObject;
-  });
-  console.log(clientProgramArray, '<-- saved array');
-
-  Clients.addClientsToProgram(clientProgramArray)
-    .then(savedArray => {
-      res.status(201).json(savedArray);
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
-// ********************************************************
-// POST /clients/program
-// ********************************************************
-router.post('/program', validTokenCheck, (req, res) => {
-  let clientProgramObject = {};
-  clientProgramObject.client_id = req.body.id;
-  clientProgramObject.program_id = req.body.program_id;
-  clientProgramObject.start_date = req.body.start_date;
-  //const coach_id = req.token.coachID;
-  console.log(clientProgramObject, '<-- saved object');
-
-  Clients.addProgramToClient(clientProgramObject)
-    .then(savedObject => {
-      res.status(201).json(savedObject);
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
 module.exports = router;
